+++ packages/api-client/src/team/conversation/index.ts
@@ -17,4 +17,5 @@
  *
  */
 
+export * from './TeamConversation';
 export * from './TeamConversationAPI';